Nexa-Studio-Full-Old-Codebase/Nexa-Studio/requirements.txt
flask==2.3.3
weasyprint==60.1
jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0 
psycopg2-binary==2.9.10 
langfuse>=2.55.3
openai>=1.0.0 